requirements.txt
python-telegram-bot
aiosqlite
ollama
whisper
torch
torchaudio
TTS
transitions
cachetools
aiolimiter
pyyaml
langdetect